+++ example/example_dart/test/example_dart_test.dart
@@ -1,8 +1,0 @@
-import 'package:example_dart/example_dart.dart';
-import 'package:test/test.dart';
-
-void main() {
-  test('calculate', () {
-    expect(calculate(), 42);
-  });
-}